+++ src/main/java/dan200/computercraft/client/render/TileEntityMonitorRenderer.java
@@ -8,21 +8,23 @@ package dan200.computercraft.client.render;
 import com.mojang.blaze3d.platform.GlStateManager;
 import com.mojang.blaze3d.platform.MemoryTracker;
 import com.mojang.blaze3d.systems.RenderSystem;
-import com.mojang.blaze3d.vertex.*;
+import com.mojang.blaze3d.vertex.PoseStack;
+import com.mojang.blaze3d.vertex.VertexConsumer;
 import com.mojang.math.Matrix3f;
 import com.mojang.math.Matrix4f;
 import com.mojang.math.Vector3f;
 import dan200.computercraft.ComputerCraft;
 import dan200.computercraft.client.FrameInfo;
-import dan200.computercraft.client.gui.FixedWidthFontRenderer;
+import dan200.computercraft.client.render.text.DirectFixedWidthFontRenderer;
+import dan200.computercraft.client.render.text.FixedWidthFontRenderer;
+import dan200.computercraft.client.util.DirectBuffers;
 import dan200.computercraft.core.terminal.Terminal;
-import dan200.computercraft.core.terminal.TextBuffer;
 import dan200.computercraft.shared.peripheral.monitor.ClientMonitor;
 import dan200.computercraft.shared.peripheral.monitor.MonitorRenderer;
 import dan200.computercraft.shared.peripheral.monitor.TileMonitor;
-import dan200.computercraft.shared.util.Colour;
 import dan200.computercraft.shared.util.DirectionUtil;
 import net.minecraft.client.renderer.MultiBufferSource;
+import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.blockentity.BlockEntityRenderer;
 import net.minecraft.client.renderer.blockentity.BlockEntityRendererProvider;
 import net.minecraft.core.BlockPos;
@@ -34,8 +36,8 @@ import org.lwjgl.opengl.GL31;
 import javax.annotation.Nonnull;
 import java.nio.ByteBuffer;
 
-import static dan200.computercraft.client.gui.FixedWidthFontRenderer.*;
-import static dan200.computercraft.client.render.RenderTypes.FULL_BRIGHT_LIGHTMAP;
+import static dan200.computercraft.client.render.text.FixedWidthFontRenderer.FONT_HEIGHT;
+import static dan200.computercraft.client.render.text.FixedWidthFontRenderer.FONT_WIDTH;
 
 public class TileEntityMonitorRenderer implements BlockEntityRenderer<TileMonitor>
 {
@@ -43,8 +45,8 @@ public class TileEntityMonitorRenderer implements BlockEntityRenderer<TileMonito
      * {@link TileMonitor#RENDER_MARGIN}, but a tiny bit of additional padding to ensure that there is no space between
      * the monitor frame and contents.
      */
-    private static final float MARGIN = (float) (TileMonitor.RENDER_MARGIN * 1.05);
-    private static ByteBuffer tboContents;
+    private static final float MARGIN = (float) (TileMonitor.RENDER_MARGIN * 1.1);
+    private static ByteBuffer backingBuffer;
 
     private static final Matrix3f IDENTITY;
 
@@ -60,7 +62,7 @@ public class TileEntityMonitorRenderer implements BlockEntityRenderer<TileMonito
     }
 
     @Override
-    public void render( @Nonnull TileMonitor monitor, float partialTicks, @Nonnull PoseStack transform, @Nonnull MultiBufferSource renderer, int lightmapCoord, int overlayLight )
+    public void render( @Nonnull TileMonitor monitor, float partialTicks, @Nonnull PoseStack transform, @Nonnull MultiBufferSource bufferSource, int lightmapCoord, int overlayLight )
     {
         // Render from the origin monitor
         ClientMonitor originTerminal = monitor.getClientMonitor();
@@ -120,32 +122,27 @@ public class TileEntityMonitorRenderer implements BlockEntityRenderer<TileMonito
             // Avoid PoseStack#scale to preserve normal matrix.
             transform.last().pose().multiply( Matrix4f.createScaleMatrix( (float) xScale, (float) -yScale, 1.0f ) );
 
-            renderTerminal( renderer, transform, originTerminal, (float) (MARGIN / xScale), (float) (MARGIN / yScale) );
-
-            // We don't draw the cursor with the VBO, as it's dynamic and so we'll end up refreshing far more than is
-            // reasonable.
-            FixedWidthFontRenderer.drawCursor(
-                transform, renderer.getBuffer( RenderTypes.MONITOR ),
-                0, 0, terminal, !originTerminal.isColour()
-            );
+            renderTerminal( bufferSource, transform, originTerminal, (float) (MARGIN / xScale), (float) (MARGIN / yScale) );
 
             transform.popPose();
         }
         else
         {
             FixedWidthFontRenderer.drawEmptyTerminal(
-                transform, renderer.getBuffer( RenderTypes.MONITOR ),
+                FixedWidthFontRenderer.toVertexConsumer( transform, bufferSource.getBuffer( RenderTypes.MONITOR ) ),
                 -MARGIN, MARGIN,
-                (float) (xSize + 2 * MARGIN), (float) -(ySize + MARGIN * 2), FULL_BRIGHT_LIGHTMAP
+                (float) (xSize + 2 * MARGIN), (float) -(ySize + MARGIN * 2)
             );
         }
 
         transform.popPose();
     }
 
-    private static void renderTerminal( @Nonnull MultiBufferSource renderer, PoseStack transform, ClientMonitor monitor, float xMargin, float yMargin )
+    private static void renderTerminal( @Nonnull MultiBufferSource bufferSource, PoseStack transform, ClientMonitor monitor, float xMargin, float yMargin )
     {
         Terminal terminal = monitor.getTerminal();
+        int width = terminal.getWidth(), height = terminal.getHeight();
+        int pixelWidth = width * FONT_WIDTH, pixelHeight = height * FONT_HEIGHT;
 
         MonitorRenderer renderType = MonitorRenderer.current();
         boolean redraw = monitor.pollTerminalChanged();
@@ -155,34 +152,15 @@ public class TileEntityMonitorRenderer implements BlockEntityRenderer<TileMonito
         {
             case TBO:
             {
-                int width = terminal.getWidth(), height = terminal.getHeight();
-
-                int pixelWidth = width * FONT_WIDTH, pixelHeight = height * FONT_HEIGHT;
                 if( redraw )
                 {
-                    int size = width * height * 3;
-                    if( tboContents == null || tboContents.capacity() < size )
-                    {
-                        tboContents = MemoryTracker.create( size );
-                    }
+                    var terminalBuffer = getBuffer( width * height * 3 );
+                    MonitorTextureBufferShader.setTerminalData( terminalBuffer, terminal );
+                    DirectBuffers.setBufferData( GL31.GL_TEXTURE_BUFFER, monitor.tboBuffer, terminalBuffer, GL20.GL_STATIC_DRAW );
 
-                    ByteBuffer monitorBuffer = tboContents;
-                    monitorBuffer.clear();
-                    for( int y = 0; y < height; y++ )
-                    {
-                        TextBuffer text = terminal.getLine( y ), textColour = terminal.getTextColourLine( y ), background = terminal.getBackgroundColourLine( y );
-                        for( int x = 0; x < width; x++ )
-                        {
-                            monitorBuffer.put( (byte) (text.charAt( x ) & 0xFF) );
-                            monitorBuffer.put( (byte) getColour( textColour.charAt( x ), Colour.WHITE ) );
-                            monitorBuffer.put( (byte) getColour( background.charAt( x ), Colour.BLACK ) );
-                        }
-                    }
-                    monitorBuffer.flip();
-
-                    GlStateManager._glBindBuffer( GL31.GL_TEXTURE_BUFFER, monitor.tboBuffer );
-                    GlStateManager._glBufferData( GL31.GL_TEXTURE_BUFFER, monitorBuffer, GL20.GL_STATIC_DRAW );
-                    GlStateManager._glBindBuffer( GL31.GL_TEXTURE_BUFFER, 0 );
+                    var uniformBuffer = getBuffer( MonitorTextureBufferShader.UNIFORM_SIZE );
+                    MonitorTextureBufferShader.setUniformData( uniformBuffer, terminal, !monitor.isColour() );
+                    DirectBuffers.setBufferData( GL31.GL_UNIFORM_BUFFER, monitor.tboUniform, uniformBuffer, GL20.GL_STATIC_DRAW );
                 }
 
                 // Nobody knows what they're doing!
@@ -192,10 +170,10 @@ public class TileEntityMonitorRenderer implements BlockEntityRenderer<TileMonito
                 RenderSystem.activeTexture( active );
 
                 MonitorTextureBufferShader shader = RenderTypes.getMonitorTextureBufferShader();
-                shader.setupUniform( width, height, terminal.getPalette(), !monitor.isColour() );
+                shader.setupUniform( monitor.tboUniform );
 
                 Matrix4f matrix = transform.last().pose();
-                VertexConsumer buffer = renderer.getBuffer( RenderTypes.MONITOR_TBO );
+                VertexConsumer buffer = bufferSource.getBuffer( RenderTypes.MONITOR_TBO );
                 tboVertex( buffer, matrix, -xMargin, -yMargin );
                 tboVertex( buffer, matrix, -xMargin, pixelHeight + yMargin );
                 tboVertex( buffer, matrix, pixelWidth + xMargin, -yMargin );
@@ -206,46 +184,76 @@ public class TileEntityMonitorRenderer implements BlockEntityRenderer<TileMonito
 
             case VBO:
             {
-                VertexBuffer vbo = monitor.buffer;
+                var vbo = monitor.buffer;
                 if( redraw )
                 {
-                    Tesselator tessellator = Tesselator.getInstance();
-                    BufferBuilder builder = tessellator.getBuilder();
-                    builder.begin( RenderTypes.MONITOR.mode(), RenderTypes.MONITOR.format() );
+                    int vertexSize = RenderTypes.MONITOR.format().getVertexSize();
+                    ByteBuffer buffer = getBuffer( DirectFixedWidthFontRenderer.getVertexCount( terminal ) * vertexSize );
 
-                    FixedWidthFontRenderer.drawTerminalWithoutCursor(
-                        new PoseStack(), builder, 0, 0,
-                        terminal, !monitor.isColour(), yMargin, yMargin, xMargin, xMargin, FULL_BRIGHT_LIGHTMAP
+                    // Draw the main terminal and store how many vertices it has.
+                    DirectFixedWidthFontRenderer.drawTerminalWithoutCursor(
+                        buffer, 0, 0, terminal, !monitor.isColour(), yMargin, yMargin, xMargin, xMargin
                     );
+                    int termIndexes = buffer.position() / vertexSize;
 
-                    builder.end();
-                    vbo.upload( builder );
+                    // If the cursor is visible, we append it to the end of our buffer. When rendering, we can either
+                    // render n or n+1 quads and so toggle the cursor on and off.
+                    DirectFixedWidthFontRenderer.drawCursor( buffer, 0, 0, terminal, !monitor.isColour() );
+
+                    buffer.flip();
+
+                    vbo.upload( termIndexes, RenderTypes.MONITOR.mode(), RenderTypes.MONITOR.format(), buffer );
                 }
 
                 Matrix3f popViewRotation = RenderSystem.getInverseViewRotationMatrix();
                 RenderSystem.setInverseViewRotationMatrix( IDENTITY );
 
+                bufferSource.getBuffer( RenderTypes.MONITOR );
                 RenderTypes.MONITOR.setupRenderState();
-                vbo.drawWithShader( transform.last().pose(), RenderSystem.getProjectionMatrix(), RenderTypes.getMonitorShader() );
-                RenderTypes.MONITOR.clearRenderState();
+
+                var matrix = transform.last().pose();
+                if( MonitorRenderer.canvasModPresent )
+                {
+                    var modelViewMatrix = RenderSystem.getModelViewMatrix().copy();
+                    modelViewMatrix.multiply( matrix );
+                    matrix = modelViewMatrix;
+                }
+
+                vbo.drawWithShader(
+                    matrix, RenderSystem.getProjectionMatrix(), RenderTypes.getMonitorShader(),
+                    // As mentioned in the above comment, render the extra cursor quad if it is visible this frame. Each
+                    // quad has an index count of 6.
+                    FixedWidthFontRenderer.isCursorVisible( terminal ) && FrameInfo.getGlobalCursorBlink() ? vbo.getIndexCount() + 6 : vbo.getIndexCount()
+                );
 
                 RenderSystem.setInverseViewRotationMatrix( popViewRotation );
-
-                // How we would draw the terminal without using VBO, for reference until I need to delete this.
-                //VertexConsumer buffer = renderer.getBuffer( RenderTypes.MONITOR );
-                //FixedWidthFontRenderer.drawTerminalWithoutCursor(
-                //    transform, buffer, 0, 0,
-                //    terminal, !monitor.isColour(), yMargin, yMargin, xMargin, xMargin, FULL_BRIGHT_LIGHTMAP
-                //);
                 break;
             }
         }
+
+        // Force a flush of the buffer. WorldRenderer.updateCameraAndRender will "finish" all the built-in buffers
+        // before calling renderer.finish, which means our TBO quad or depth blocker won't be rendered yet!
+        bufferSource.getBuffer( RenderType.solid() );
     }
 
     private static void tboVertex( VertexConsumer builder, Matrix4f matrix, float x, float y )
     {
         // We encode position in the UV, as that's not transformed by the matrix.
         builder.vertex( matrix, x, y, 0 ).uv( x, y ).endVertex();
+    }
+
+    @Nonnull
+    private static ByteBuffer getBuffer( int capacity )
+    {
+
+        ByteBuffer buffer = backingBuffer;
+        if( buffer == null || buffer.capacity() < capacity )
+        {
+            buffer = backingBuffer = buffer == null ? MemoryTracker.create( capacity ) : MemoryTracker.resize( buffer, capacity );
+        }
+
+        buffer.clear();
+        return buffer;
     }
 
     @Override