+++ src/main/java/dan200/computercraft/shared/peripheral/generic/GenericPeripheral.java
@@ -12,9 +12,9 @@ import dan200.computercraft.api.lua.MethodResult;
 import dan200.computercraft.api.peripheral.IComputerAccess;
 import dan200.computercraft.api.peripheral.IDynamicPeripheral;
 import dan200.computercraft.api.peripheral.IPeripheral;
+import net.minecraft.core.Registry;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.level.block.entity.BlockEntity;
-import net.minecraft.core.Registry;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;