+++ src/main/java/dan200/computercraft/shared/turtle/blocks/TileTurtle.java
@@ -432,10 +432,13 @@ public class TileTurtle extends TileComputerBase implements ITurtleTile, Default
     @Override
     public void setItem( int i, @Nonnull ItemStack stack )
     {
-        if( i >= 0 && i < INVENTORY_SIZE && !InventoryUtil.areItemsEqual( stack, inventory.get( i ) ) )
+        if ( i >= 0 && i < INVENTORY_SIZE )
         {
             inventory.set( i, stack );
-            onInventoryDefinitelyChanged();
+            if ( !InventoryUtil.areItemsEqual( stack, inventory.get( i ) ) )
+            {
+                onInventoryDefinitelyChanged();
+            }
         }
     }