+++ src/main/java/dan200/computercraft/shared/turtle/upgrades/TurtleSpeaker.java
@@ -11,14 +11,13 @@ import dan200.computercraft.api.turtle.AbstractTurtleUpgrade;
 import dan200.computercraft.api.turtle.ITurtleAccess;
 import dan200.computercraft.api.turtle.TurtleSide;
 import dan200.computercraft.api.turtle.TurtleUpgradeType;
+import dan200.computercraft.shared.peripheral.speaker.SpeakerPosition;
 import dan200.computercraft.shared.peripheral.speaker.UpgradeSpeakerPeripheral;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.client.resources.model.ModelResourceLocation;
-import net.minecraft.core.BlockPos;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.level.Level;
 import net.minecraft.world.phys.Vec3;
 
 import javax.annotation.Nonnull;
@@ -41,18 +40,11 @@ public class TurtleSpeaker extends AbstractTurtleUpgrade
             this.turtle = turtle;
         }
 
-        @Override
-        public Level getLevel()
-        {
-            return turtle.getLevel();
-        }
-
         @Nonnull
         @Override
-        public Vec3 getPosition()
+        public SpeakerPosition getPosition()
         {
-            BlockPos pos = turtle.getPosition();
-            return new Vec3( pos.getX() + 0.5, pos.getY() + 0.5, pos.getZ() + 0.5 );
+            return SpeakerPosition.of( turtle.getLevel(), Vec3.atCenterOf( turtle.getPosition() ) );
         }
 
         @Override