+++ src/main/resources/data/computercraft/lua/rom/help/whatsnew.md
@@ -1,7 +1,9 @@
-New features in CC: Tweaked 1.100.2
+New features in CC: Tweaked 1.100.5 (CC: Restitched 1.100.4)
+
+* Improve performance of monitor rendering.
 
 Several bug fixes:
-* Fix wired modems swapping the modem/peripheral block state.
-* Remove debugging logging line from `turtle.attack`.
+* Various documentation fixes (bclindner, Hasaabitt)
+* Speaker sounds are now correctly positioned on the centre of the speaker block.
 
 Type "help changelog" to see the full version history.